Conflicts:
	Babylon/Materials/textures/babylon.texture.js
	Babylon/Materials/textures/babylon.texture.ts
	Babylon/babylon.engine.js
	Babylon/babylon.engine.ts

Former-commit-id: d54cb710e8bb62579a6ad3bca9a2d3814f081a6b

+++ Babylon/Shaders/color.vertex.fx
@@ -1,4 +1,4 @@
-﻿precision mediump float;
+﻿precision highp float;
 
 // Attributes
 attribute vec3 position;
